requirements.txt
torch
torchvision
numpy
matplotlib
scikit-image
scipy
pyyaml
pillow
optuna
azureml-core
